requirements.txt
# NIE Tamil AI Tutor — Enterprise Architecture
# ============================================

# --- Core LLM backends ---
google-genai>=1.0.0
python-dotenv>=1.0.0

# --- Vector DB & Embeddings ---
chromadb>=0.5.0
sentence-transformers>=3.0.0

# --- PDF processing ---
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0

# --- API server ---
fastapi>=0.115.0
uvicorn>=0.30.0

# --- Optional: Ollama local fallback ---
ollama>=0.2.0
httpx>=0.27.0

# --- Utilities ---
numpy>=1.26.0
tqdm>=4.66.0
